Updating styleguide config, using composition
This PR has been set up so that the styleguide's configuration is a
little less daunting to utilize. This change allows for devs to create
components that will be added to the Pattern Library without having to
go into the config file and add a component each time they create a new
component.

I have also modified the way that I have built some of the components
moving away from props to composition instead. An example of this is the
buttons. Instead of having a `<Button/>` component with a prop for a
theme, I have instead created different buttons for the various styles
that they have (e.g. `<Button />, <WhiteBorderButton />
<BlueBorderButton/>`). I also included the concept of a `.base.js`
component. These components shouldn't be used on their own however, are
used as a base line to compose final components. An example of this is
the `Button.base.js` component. It defines icons that appear in the
buttons, some of the states such as loading, disabled, selected, and
checked however, the visual touches are finalized in the components that
comprise the base component (e.g. <Button />, <BlueBorderButton />,
<WhiteBorderButton />).