Moving more components to composition based styles
After working on refactoring the buttons in the Pattern Library to make
better use of composition, I have taken the liberty to convert more
components to a composed architechture with `.base` files to modify.

While there is a larger library of actual components, my gut feeling is
that they are easier to utilize in the long run.