Upgrading our node_modules and new Components
There was an issue where the build system had broken for me, so I went
about updating the node modules for this project. Doing this also caused
and issue where some existing functionality from styled components no
longer worked. That required a bit of tweaking there, but mainly
involved replacing injectGlobalStyle with createGlobalStyle and an issue
with keyframes specifically in the Chevron component.

I updated the Image component to no longer have the fade in animation
(it wasn't working anyways) and I added the ability to provide a srcSet
and sizes props to it in a means that I hope is verbose enough by
providing an object with the size/width as the key and then providing
the source or the mediaquery as the value.

I also included the brand spanking new video component! I had originally
tried to have cloudinary automatically generate different file types
when a user would simply provide a single file source. This however
didn't seem to work as I was expecting, it appeared that cloudinary
would simply retain the original file type for the video when I tried to
supply different formats. What I instead did was to make the sources
prop be more flexible by either providing a string or array value. In
doing so the code now automatically generates the source elements with
the correct format based on the file extension. We can alternately
simply provide `<source />` elements as children of the `Video`
component if we so choose.