Adding the new BackgroundImage component
This PR adds the new BackgroundImage component. This component will
pick out the correctly sized background image based on the current width
of the element and set it's size based on it's aspect ratio! I also had
to rename the Image component because *fun fact* it was overriding the
default javascript Image function.