Fixing an image load issue with background image
There was an issue when updating the image src would not load it in
immediately. I changed where we actually render the `Image` element,
which should also improve performance as it was firing during the render
function each time where now it should only fire when the state is
updated.