Updating tests to pull from SRC directory
As discussed with the issues of components not being exported correctly,
I have updated the tests to pull from the root `index.js` file located
in the top of the `src` directory. We should try to adhere to this
pattern as tests will blow up if the component that you are trying to
test is not being exported up in the top index.js file.