Adding the Customer Quotes component
This PR adds the new CustomerQuotes component to the Component Library.
It takes in an array of quotes and then sets a timer that it uses to
render out a new quote every five seconds. It the component reaches the
end of the quotes array, then it will loop back around.