Making BackgroundVideo responsive
Due to an issue where our content was taller than our background video
we had some issues where the children of the tout were overflowing in
the Tout. This PR allows us to have two video files that are pulled in
when a specific breakpoint is reached. This paradigm is simmilar to how
we do backgroundImages in Mirage.