Adding screen reader region for accessibility
This PR adds a reagion that renders out all of the press quotes and
hides the quotes that we render in and out from the screen reader. This
allows users with screen readers to navigate through the quotes without
any strange behavior due to the quotes being mounted and unmounted.