Tweaking FlexRow and FlexCol
In an effort to help clean up the DOM Structure, I switched FlewRow and
FlexCol to simplay apply the styles to the elements themselves. They are
given a default of `div` but can be passed a string element such as
`nav`, `img` and even React Components themselves. My hope is that this
should eliminate a lot of wrapper components.