Making Breakpoint Component SSR Friendly
Based off of a comment that @jrusso1020 had made in the breakpoint
components, we went back and converted the Breakpoint components to
utilize CSS instead of React Responsive. The components should hopefully
have the same functionality however, they should be able to work even if
they are server side rendered!