Introducing Contentful to Mirage
This PR introduces Contentful as a development dependency to the Mirage
codebase. We also created a Contentful React component that allows
developers to make a call to Contentful to then get real data back from
the contentful service. Since Contentful is installed as a development
dependency it shouldn't be included when installing a later version of
Mirage thus creating conflicting version of Contentful. The Contenful
component itself is also not being exported as a component currently, if
we decide that it might be easier to utilize it as component then we
could include it in the main index.js export of Mirage.

This PR also includes the first example using the Contentful component
the ContentfulImage component. We pass the operation to `getAsset` and
the `id` of the image asset from:
https://app.contentful.com/spaces/efh696jqykjs/assets/VY7BmW2jtHaLTjQEQKe3t
and it then gets the response data from Contentful and properly passes
it down to an `InlineImage` component.