Inline Images can be Lazy Loaded
[#167436130]

This PR adds the ability to lazy load images. If the `lazyLoad` prop is passed to the `InlineImage` we will not set the `src` attribute on the `img` element but will set a `data-src` atribute which can then be interacted with a HOC to load in images if the element is in the current viewport.