Update progress bar logic and copy
With this change we'll update the logic used to determine
the fill on the progress bar inside the SideCart, so we track
the evergreen promotion (Get 20% off with 4+ items) instead of
the free shipping promotion.

We'll also make minor changes to the sidebar layout and update
the copy above the progress bar so customers know how close they
are to getting the discount.

This change also updates a lot of outdated snapshots and attempts
to remove a few warnings throughout mirage.