Fix copy and styling across the cart sidebar
With this change we'll add some copy to the bottom of the
cart sidebar to let customers know they'll get free shipping
on orders over $50.

We'll also update the copy used in the ProgressBarText to better
reflect the evergreen promo savings that the customer will get.

We'll also fix a bug were adding 4+ items to the bag introduced
horizontal scrolling on the sidebar.

*Bonus* we'll add the particleSpeed prop to the Confetti component
so we can control how fast confetti particles fall down.