Allow mobile customers to view quick view with a single tap
With this change we'll allow customers browsing our site on
their mobile devices to access quick view by tapping on the
slider of each product tile. Desktop users will only be
able to access quick view by clicking on the quick view button
which will only be visible in desktop.