Add necessary components to build quick view
With this change we'll migrate a set of components from thor
into mirage so we can build the quick view component on
quark-ui. The idea is to share as much logic and styling
as possible between thor and quark-ui.

A QuickView component hasn't been exported because some of the
components inside of it need to be connected to the state,
so instead of passing a bunch of properties to the parent
component, we opted for exporting smaller presentational components
and building the QuickView component on each app, with minimal
duplicate logic.