Ensure that images on the ProductTile slider match the current colorway
With this change we'll make sure that images in the ProductTile
slider match the currently selected colorway. Basically the image
wouldn't change before because we were spreading all of the
props from ProductTile and passing them to the Slider, which
caused the lazyLoad prop to override the value of the lazyLoad
key coming from state.

Since lazyLoad was true, we were attempting to lazy load the
images for the newly selected colorway, which would only happen
if the user scrolled. With this change we'll stop spreading
props and passing them to the Slider which should make it obvious
which properties it's receiving at any point in time.