Adding mini theme to Mirage (#326)
This PR establishes a theme for mini that can be used to automatically style all mirage components scoped under the <ROATheme theme='mini' /> ThemeProvider. Previously, we had been adding a 'kind' prop to certain components that needed a different set of CSS rules for mini. Ultimately though, this became difficult to maintain and limited our ability to reuse composed components (ie AddressFields or PaymentFields in Thor) in non-mini contexts.

For now, the only differences between the two themes will be their color schemes and border styles on button. With this infrastructure in place, making additional stylistic changes will be much more straightforward.

Unfortunately, react-styleguidist does not have support for accessing the context of a Wrapper component from within the Pattern demo library. This means we are unable to view the mini-themed components with our default settings. It looks like there's some discussion on adding this support, and an effort at implementing it that has since been abandoned but nothing exists at the moment. To view mini-specific styles, I've been resorting to changing the value of the default themeName prop in ROATheme to "mini." It's frustrating, but our options seem kind of limited here from what I can tell.

My alternative idea was to wrap each example of a mini-themed component in a , but this would require us to import the miniTheme object and the ThemeProvider itself into our example markdown files. Our version of react-styleguidist does not support the inclusion of import statements in the js blocks of the example files however. We'd need to upgrade this package to v9 to gain access to this functionality, and this would in turn require us to upgrade Webpack from v3 -> v4. 😬 It would probably be good to do this at some point, so let me know what you think!