Use relative paths for navigation
With this change we'll update the desktop and mobile navigation
components so they use relative paths instead of absolute paths,
this makes links protocol agnostic which should make the links
more flexible.

We'll also export the sortShots function as it is used in a ton
of different places (PLP, PDP, product tiles, etc.) being able
to import this function will help reduce duplication.