Fix flaky test suite (#335)
This PR addresses the long-standing issue we've had with flaky tests in our Mirage test suite. Although the error messages were consistently vague, and seemingly unrelated to the tests that Jest claimed were failing, I think our problems were being cause by a race condition somewhere. To address this, I modified the test command that we run in Codeship to include the --runInBand flag to ensure that the tests are run in sequence on a single process instead of being split up amongst a group of workers. This also has the potential to boost the speed of our tests running in CI!

Additionally, I updated and committed all snapshots that were previously being ignored to conform with Jest's recommendation. Finally, I addressed a number of warnings that our tests were generating to reduce the noisiness of our tests logs should we need to read them in the future. These changes mostly involved adding required props to our test instances that were not being provided.