Update Product Sku in AddToBagState Component When Client-Side Navigating (#338)
This PR addresses a bug in which clicking "Add to Bag" on PDPs loaded via client-side nagivating would duplicate the last item added to your cart. Because sizeSelected and sku are treated as internal pieces of state in this component, their old values would persist even when the selectedColorway changed. By resetting these state values when selectedColorway changes, we ensure that the previous item is not added to their cart again.