Simplify the Display of Subscription-Related Links in Nav (#340)
This PR simplifies the way that we conditionally display subscription links in the nav. Because the isNotSubscriber and isSubscriptionMember props are exact opposites of each other, we could just negate the value of one when we want to show the links associated with the other. As a result, isNotSubsriber was removed in favor of the renamed isSubscriptionMember/showSubscriptionLinks. Additionally, a spinner has been added to the account links dropdown anchor that will appear in the time it takes Redux to populate a user's name in store. Once this value has been fetched, the spinner will be replaced by the user's name.